Modularizei o modo clássico
