+++ Modo-Clássico/index.html
@@ -31,6 +31,6 @@
 
       <audio id="bg-music" src="music.mp3" loop></audio>
       
-      <script src="script.js"></script>
+      <script type="module" src="js/main.js"></script>
 </body>
 </html>
