+++ Modo-Clássico/js/config.js
@@ -1,0 +1,38 @@
+
+export const canvas = document.querySelector("#space-invaders");
+export const ctx = canvas.getContext("2d");
+export const playBtn = document.querySelector("#play-btn");
+export const menu = document.querySelector("#menu");
+export const muteBtn = document.querySelector("#mute-btn")
+
+export const playerImg = new Image();
+playerImg.src = "assets/SpaceShip(192x192)_0001.png";
+export const playerImg_frame2 = new Image();
+playerImg_frame2.src = "assets/SpaceShip(192x192)_0002.png"
+export const playerImg_frame3 = new Image();
+playerImg_frame3.src = "assets/SpaceShip(192x192)_0003.png"
+export const playerFrames = [playerImg, playerImg_frame2, playerImg_frame3]
+export const enemyImg1 = new Image();
+enemyImg1.src = "assets/Alien1(192x192).png";
+export const enemyImg1_frame2 = new Image();
+enemyImg1_frame2.src = "assets/Alien1-Quadro2(192x192).png";
+export const enemyImg2 = new Image();
+enemyImg2.src = "assets/Alien2(192x192)_0001.png";
+export const enemyImg2_frame2 = new Image();
+enemyImg2_frame2.src = "assets/Alien2(192x192)_0002.png";
+export const enemyImg3 = new Image();
+enemyImg3.src = "assets/Alien3(192x192)_0001.png";
+export const enemyImg3_frame2 = new Image();
+enemyImg3_frame2.src = "assets/Alien3(192x192)_0002.png";
+export const baseImg = new Image();
+baseImg.src = "assets/escudo(192x192).png";
+export const vidaImg = new Image();
+vidaImg.src = "assets/vida(192x192).png";
+export const semvidaImg = new Image();
+semvidaImg.src = "assets/sem-vida(192x192).png";
+export const playerShotSound = new Audio();
+playerShotSound.src = "assets/tiro-nave.mp3";
+playerShotSound.volume = 0.40
+export const baseDestroyedSound = new Audio();
+baseDestroyedSound.src = "assets/explosao.mp3"
+baseDestroyedSound.volume = 0.6
